requirements.txt
cycler==0.11.0
decorator==4.4.2
imageio==2.15.0
joblib==1.1.1
kiwisolver==1.3.1
matplotlib==3.3.4
networkx==2.5.1
numpy==1.19.4
opencv-python==4.1.1.26
Pillow==8.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyWavelets==1.1.1
scikit-image==0.17.2
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
threadpoolctl==3.1.0
tifffile==2020.9.3